requirements.txt
flask>=2.0.0
flask-socketio>=5.1.1
python-socketio>=5.4.0
eventlet>=0.30.2
gevent>=22.10.2
gevent-websocket>=0.10.1
websockets>=10.0
requests>=2.26.0
rich>=10.0.0
psutil>=5.8.0
python-engineio>=4.0.0
aiohttp>=3.8.0
python-dotenv>=0.19.0
pyyaml>=5.4.1
groq>=0.3.0
ollama>=0.1.5
dbus-python>=1.2.18
pillow>=9.0.0
gTTS>=2.2.4
SpeechRecognition>=3.8.1
pyttsx3>=2.90
pvporcupine>=2.1.0
sounddevice>=0.4.4
numpy>=1.21.0
scipy>=1.7.0
watchdog>=2.1.0
beautifulsoup4>=4.9.3
PyGObject>=3.42.0
pyaudio>=0.2.11 